Initial Documentation of the Setup
The README.md provides a brief layout why this example exists. It
then furtermore explains how to setup the devdataserver, how to setup
the frontend and how to use the frontend to display data from a chosen
device.

Note: The Code is documented in the code itself via comments that show
different steps taken to achieve the example result.