Feature: Integration of multiple Datasources
Issue: With the current code design only one datasource was possible.
Adding another datasource would result in an unscalable mess of code .

Therefor a design pattern was chosen that allows adding multiple
datasources that implement the same DataSourceStrategy.
The plot and source selection is now more independent from the datasource
that is chosen .

This made it easy to integrate also the omnaiscope dataserver as a
second example, providing a real life datasource that measures physical
data.