Feature: Recording a measurement
Visualizing a measurement is possible but not recording one.
Record means: Saving a certain timerange of the measurement.

Changes:
- Record button to start/stop recording
- Record Handling functions to start/stop the recording

Important: record_data is saved in the RAM as long as the measurement is
recorded. Everything is flushed at once.
Make sure to not fill up your RAM

Note: The UNIX timestamps are currently not saved correctly as they are
interpreted as floats

Userflow:
- start the measurement
- press record
- stop record
- The file should be in the main directory of this repository