Feature: Analyze mean of data
This is an example how the data measured could be analysed.

The analyze function loads the latest file that was saved.
It sends the data to the endpoint http://127.0.0.1:8000/mean which is in
this case the endpoint to the test analyse server for omnaiview.
If successful the mean value of the y-values is returned which is shown in a popup.